Changed log folder
Changed log folder to appdata, allowing logs to be created even if work folder is not set, which is critical for error log.

Also allows the work folder to be kept clean for sharing.

If separation of logs by project would be necessary, it can be done in teh fuuture in the new folder.